pyproject.toml
[project]
name = "trash_classifier"
version = "0.1.0"
description = ""
authors = [
    {name = "Thanh Tran",email = "tnt04052001@gmail.com"}
]
readme = "README.md"
[tool.poetry.dependencies]
python = "<3.13,>=3.9"
torch = "^2.1.0"
torchvision = "^0.16.0"
pandas = "^2.1.1"
numpy = "^1.26.0"
hydra-core = "^1.3.2"
tdqm = "^0.0.1"
torchsummary = "^1.5.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
pytest = "^7.4.2"
flake8 = "^6.1.0"
isort = "^5.12.0"
pre-commit = "^3.4.0"
dvc = "^3.30.3"
dvc-gdrive = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 90
target-version = ["py38"]

[tool.isort]
line_length = 90
lines_after_imports = 2

[tool.flake8]
max-line-length = 90
exclude = [".git", "__pycache__"]
max-complexity = 10